+++ ouds_theme_contract/lib/theme/tokens/components/ouds_components_tokens.dart
@@ -11,6 +11,7 @@
  */
 
 import 'package:ouds_theme_contract/ouds_tokens_provider.dart';
+import 'package:ouds_theme_contract/theme/tokens/components/ouds_buttonMono_tokens.dart';
 import 'package:ouds_theme_contract/theme/tokens/components/ouds_button_tokens.dart';
 import 'package:ouds_theme_contract/theme/tokens/components/ouds_checkbox_tokens.dart';
 import 'package:ouds_theme_contract/theme/tokens/components/ouds_controlItem_tokens.dart';
@@ -23,6 +24,7 @@ import 'ouds_chip_tokens.dart';
 
 class OudsComponentsTokens {
   final OudsButtonTokens button;
+  final OudsButtonMonoTokens buttonMono;
   final OudsCheckboxTokens checkbox;
   final OudsControlItemTokens controlItem;
   final OudsRadioButtonTokens radioButton;
@@ -34,6 +36,7 @@ class OudsComponentsTokens {
   OudsComponentsTokens({
     required OudsProvidersTokens providersTokens,
     OudsButtonTokens? button,
+    OudsButtonMonoTokens? buttonMono,
     OudsCheckboxTokens? checkbox,
     OudsControlItemTokens? controlItem,
     OudsRadioButtonTokens? radioButton,
@@ -42,6 +45,7 @@ class OudsComponentsTokens {
     OudsSwitchTokens? switchButton,
     OudsChipTokens? chip,
   })  : button = button ?? OudsButtonTokens(providersTokens: providersTokens),
+        buttonMono = buttonMono ?? OudsButtonMonoTokens(providersTokens: providersTokens),
         radioButton = radioButton ?? OudsRadioButtonTokens(providersTokens: providersTokens),
         checkbox = checkbox ?? OudsCheckboxTokens(providersTokens: providersTokens),
         controlItem = controlItem ?? OudsControlItemTokens(providersTokens: providersTokens),